Ensure the 'bbp_reply_to' value is set when clicking on Quote button.
Previously, we did not set the reply to post ID when the Quote button
was clicked. This meant that those viewing replies in threaded mode
would not see responses properly threaded.

This commit sets the 'bbp_reply_to' hidden input field and ensures
this works in both threaded and linear display modes. Bonus is we now
piggyback off of bbPress' reply JS in threaded mode.

Fixes #2.